Move variant analysis header tests to individual components
